requirements.txt
Cython==0.29.35
pysubs2==1.6.0
python-dotenv==1.0.0
